wishlist/requirements.txt
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
Django==2.2
django-storages==1.10.1
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gunicorn==20.0.0
idna==2.10
Pillow==8.0.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
requests==2.25.0
rsa==4.6
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
